Minor fixes after review.
