+++ .flake8
@@ -1,2 +1,8 @@
 [flake8]
+show-source=true
+statistics=true
 max-line-length = 80
+
+exclude =
+  .git,
+  **/data/**
