+++ .gitignore
@@ -1,1 +1,3 @@
 data
+__pycache__
+path.sh
