+++ egs/librispeech/ASR/local/compute_fbank_musan.py
@@ -7,43 +7,12 @@ and generated fbank features are saved in data/fbank.
 """
 
 import os
-import subprocess
-from contextlib import contextmanager
 from pathlib import Path
 
 from lhotse import CutSet, Fbank, FbankConfig, LilcomHdf5Writer, combine
 from lhotse.recipes.utils import read_manifests_if_cached
 
-
-@contextmanager
-def get_executor():
-    # We'll either return a process pool or a distributed worker pool.
-    # Note that this has to be a context manager because we might use multiple
-    # context manager ("with" clauses) inside, and this way everything will
-    # free up the resources at the right time.
-    try:
-        # If this is executed on the CLSP grid, we will try to use the
-        # Grid Engine to distribute the tasks.
-        # Other clusters can also benefit from that, provided a cluster-specific wrapper.
-        # (see https://github.com/pzelasko/plz for reference)
-        #
-        # The following must be installed:
-        # $ pip install dask distributed
-        # $ pip install git+https://github.com/pzelasko/plz
-        name = subprocess.check_output("hostname -f", shell=True, text=True)
-        if name.strip().endswith(".clsp.jhu.edu"):
-            import plz
-            from distributed import Client
-
-            with plz.setup_cluster() as cluster:
-                cluster.scale(80)
-                yield Client(cluster)
-            return
-    except:
-        pass
-    # No need to return anything - compute_and_store_features
-    # will just instantiate the pool itself.
-    yield None
+from icefall.utils import get_executor
 
 
 def compute_fbank_musan():
